+++ crates/webidl/src/generator.rs
@@ -1,4 +1,4 @@
-use crate::util::{leading_colon_path_ty, raw_ident, rust_ident};
+use crate::util::{camel_case_ident, leading_colon_path_ty, raw_ident, rust_ident};
 use proc_macro2::Literal;
 use proc_macro2::TokenStream;
 use quote::format_ident;
@@ -922,6 +922,117 @@ impl Dictionary {
     }
 }
 
+pub enum NamespaceAttributeKind {
+    Getter,
+    /// Note: Per the WebIDL spec, namespace attributes are always readonly,
+    /// so this variant is currently unused. It's kept for potential future use
+    /// or non-standard WebIDL extensions.
+    #[allow(dead_code)]
+    Setter,
+}
+
+pub struct NamespaceAttribute {
+    pub js_name: String,
+    pub rust_name: String,
+    pub ty: Type,
+    pub catch: bool,
+    pub kind: NamespaceAttributeKind,
+    pub unstable: bool,
+}
+
+impl NamespaceAttribute {
+    fn generate(
+        &self,
+        options: &Options,
+        parent_name: &Ident,
+        ns_type_name: &Ident,
+        parent_js_name: &str,
+    ) -> TokenStream {
+        let NamespaceAttribute {
+            js_name,
+            rust_name,
+            ty,
+            catch,
+            kind,
+            unstable,
+        } = self;
+
+        let unstable_attr = maybe_unstable_attr(*unstable);
+        let unstable_docs = maybe_unstable_docs(*unstable);
+
+        let mdn_docs = mdn_doc(parent_js_name, Some(js_name));
+
+        let mut features = BTreeSet::new();
+
+        add_features(&mut features, ty);
+        features.remove(&parent_name.to_string());
+
+        let cfg_features = get_cfg_features(options, &features);
+
+        features.insert(parent_name.to_string());
+
+        let doc_comment = required_doc_string(options, &features);
+
+        let (prefix, attr, def) = match kind {
+            NamespaceAttributeKind::Getter => {
+                let rust_name = rust_ident(rust_name);
+
+                let ty = if *catch {
+                    quote!( Result<#ty, JsValue> )
+                } else {
+                    quote!( #ty )
+                };
+
+                (
+                    "Getter",
+                    quote!(getter,),
+                    quote!( pub fn #rust_name() -> #ty; ),
+                )
+            }
+
+            NamespaceAttributeKind::Setter => {
+                let rust_name = rust_ident(rust_name);
+
+                let ret_ty = if *catch {
+                    Some(quote!( -> Result<(), JsValue> ))
+                } else {
+                    None
+                };
+
+                (
+                    "Setter",
+                    quote!(setter,),
+                    quote!( pub fn #rust_name(value: #ty) #ret_ty; ),
+                )
+            }
+        };
+
+        let catch = if *catch { Some(quote!(catch,)) } else { None };
+
+        let doc_comment = comment(
+            format!("{prefix} for the `{parent_js_name}.{js_name}` field.\n\n{mdn_docs}"),
+            &doc_comment,
+        );
+
+        let js_name_ident = raw_ident(js_name);
+
+        quote! {
+            #unstable_attr
+            #cfg_features
+            #[wasm_bindgen(
+                #catch
+                static_method_of = #ns_type_name,
+                js_class = #parent_js_name,
+                #attr
+                js_name = #js_name_ident
+            )]
+            #doc_comment
+            #unstable_docs
+            #def
+        }
+    }
+}
+
 pub struct Function<'a> {
     pub name: Ident,
     pub js_name: String,
@@ -1015,6 +1126,7 @@ pub struct Namespace<'a> {
     pub name: Ident,
     pub js_name: String,
     pub consts: Vec<Const>,
+    pub attributes: Vec<NamespaceAttribute>,
     pub functions: Vec<Function<'a>>,
     pub unstable: bool,
 }
@@ -1025,6 +1137,7 @@ impl Namespace<'_> {
             name,
             js_name,
             consts,
+            attributes,
             functions,
             unstable,
         } = self;
@@ -1037,12 +1150,36 @@ impl Namespace<'_> {
             .map(|x| x.generate(options, name, js_name.to_string()))
             .collect::<Vec<_>>();
 
-        let functions = if functions.is_empty() {
+        // For namespace attributes, we need a type binding that represents the namespace
+        // so we can use static_method_of to access properties on it
+        let ns_type_name = rust_ident(&format!("JsNamespace{}", camel_case_ident(js_name)));
+        let js_namespace = raw_ident(js_name);
+
+        let attributes = attributes
+            .iter()
+            .map(|x| x.generate(options, name, &ns_type_name, js_name))
+            .collect::<Vec<_>>();
+
+        // Only generate the namespace type if we have attributes that need it
+        let ns_type_binding = if attributes.is_empty() {
             None
         } else {
             Some(quote! {
                 #[wasm_bindgen]
                 extern "C" {
+                    #[wasm_bindgen(js_name = #js_namespace)]
+                    pub type #ns_type_name;
+                }
+            })
+        };
+
+        let extern_block = if functions.is_empty() && attributes.is_empty() {
+            None
+        } else {
+            Some(quote! {
+                #[wasm_bindgen]
+                extern "C" {
+                    #(#attributes)*
                     #(#functions)*
                 }
             })
@@ -1064,7 +1201,9 @@ impl Namespace<'_> {
 
                 #(#consts)*
 
-                #functions
+                #ns_type_binding
+
+                #extern_block
             }
         }
     }